[refactor] auth 환경 셋팅 및 불필요 로직 제거
