+++ auth/src/main/java/com/spring/cloud/client/auth/application/dto/UserDTO.java
@@ -1,7 +1,6 @@
 package com.spring.cloud.client.auth.application.dto;
 
 public record UserDTO (
-        String username,
         String email,
         String password,
         String role
