+++ auth/src/main/java/com/spring/cloud/client/auth/infrastructure/jwt/JwtValidator.java
@@ -26,8 +26,8 @@ public class JwtValidator {
                 .getPayload();
     }
 
-    public String getUsername(String token) {
-        return parseClaims(token).get("username", String.class);
+    public String getEmail(String token) {
+        return parseClaims(token).get("email", String.class);
     }
 
     public String getRole(String token) {
